move executeFastCommand into context.go
